REEEEEEEEEEEE world space

+++ SS14.Shared/Interfaces/Map/IMapGrid.cs
@@ -23,7 +23,7 @@ namespace SS14.Shared.Interfaces.Map
         /// <summary>
         ///     The distance between the snap grid, between each center snap and between each offset snap grid location
         /// </summary>
-        int SnapSize { get; }
+        float SnapSize { get; }
 
         /// <summary>
         ///     The origin of the grid in world coordinates. Make sure to set this!
